move search methods to searchService
